pyproject.toml
[project]
name = "mirrorkey"
version = "v0.0.1"
description = "A self-hosted, pluggable mirror for secret stores providing a consistent API & offline access for your secrets."
authors = [
    {name = "ripple",email = "github@ripplefcl.com>, tigattack <?"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
